requirements.txt
# Core libraries
streamlit==1.35.0
pandas==2.1.2
numpy==1.26.1

# Data visualization
plotly==5.18.0
matplotlib==3.8.2

# API and data handling
fredapi==0.5.1
requests==2.31.0

# Environment and configuration
python-dotenv==1.0.0

# Date handling
python-dateutil==2.8.2

# Logging enhancements
colorlog==6.7.0